refactor: removed code duplicate

+++ src/app/controllers/admin-controller.js
@@ -3,11 +3,8 @@ const express = require('express');
 const router = express.Router();
 
 const slug = require('slugify');
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
-const crypto = require('crypto');
 const cache = require('../../lib/redis');
-const adminMiddleware = require('../middlewares/adminAuth');
+const userMiddleware = require('../middlewares/userAuth');
 
 const {
     courses,
@@ -18,21 +15,21 @@ const {
     materials,
 } = require('../models');
 
-function generateTokenAdmin(params) {
-    return jwt.sign({ params }, process.env.JWT_KEY_ADMIN, {
-        expiresIn: 86400,
-    });
-}
-
 /*
 /* Rotas Relacionadads a administração de usuarios
 /* Listar usuarios
 /* Deletar usuarios
 */
-router.get('/users', adminMiddleware, async (req, res) => {
+
+router.get('/users', userMiddleware, async (req, res) => {
     try {
+        const { admin } = req;
         const params = 'users:all';
         const cached = await cache.get(params);
+
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
 
         if (cached) {
             return res.status(200).json(cached);
@@ -45,9 +42,13 @@ router.get('/users', adminMiddleware, async (req, res) => {
     }
 });
 
-router.delete('/delete/users', adminMiddleware, async (req, res) => {
+router.delete('/delete/users', userMiddleware, async (req, res) => {
     try {
         const { id } = req.body;
+        const { admin } = req;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await users.destroy({ where: { id } });
         return res.status(200).json({ success: 'deletado com sucesso' });
     } catch (err) {
@@ -61,47 +62,19 @@ router.delete('/delete/users', adminMiddleware, async (req, res) => {
 /* atualizar informações
 /* atualizar senha de acesso
 /* atualizar perfil
+/* OBS: Essa rota pode bloquear ou modificar qualquer informações do usuarios desde que seja um admin
 */
 
-router.post('/signin', async (req, res) => {
+router.put('/update/users/:id', userMiddleware, async (req, res) => {
     try {
-        const { email, password } = req.body;
-        const params = `user:admin:${email}`;
+        const { admin } = req;
+        const { id } = req.params;
 
-        const cached = await cache.get(params);
-
-        if (cached) {
-            return res.status(200).json(cached);
-        }
-        const user = await users.findOne({ where: { email } });
-
-        if (!user) {
-            return res.status(400).json({ error: 'Email não cadastrado' });
-        }
-        if (user.email !== process.env.EMAIL_ADMIN) {
-            return res.status(400).json({ error: 'você não tem permissão' });
-        }
-        if (!(await bcrypt.compare(password, user.password))) {
-            return res.status(400).json({ error: 'Email e/ou senha errado' });
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
         }
 
-        const hash = crypto.randomBytes(20).toString('hex');
-
-        await users.update({ private: hash }, { where: { id: user.id } });
-        cache.set(params, user, 60 * 20);
-        return res.status(200).json({
-            admin: true,
-            token: generateTokenAdmin({ id: user.id }),
-            user,
-        });
-    } catch (err) {
-        return res.status(400).json({ err });
-    }
-});
-
-router.put('/update', adminMiddleware, async (req, res) => {
-    try {
-        await users.update({ ...req.body }, { where: { id: req.userId } });
+        await users.update({ ...req.body }, { where: { id } });
         return res.status(200).json({ sucess: 'atualizado com sucesso!' });
     } catch (err) {
         return res.status(400).json({ err });
@@ -116,49 +89,21 @@ router.put('/update', adminMiddleware, async (req, res) => {
 /* 4º Deletar Curso
 */
 
-router.get('/course/:page', adminMiddleware, async (req, res) => {
+router.post('/create/course', userMiddleware, async (req, res) => {
     try {
-        const { page } = req.params;
-        const params = `page:course:admin:${page}`;
+        const { admin } = req;
+        const { title, description, image, time, qtDisciplines } = req.body;
 
-        const cached = await cache.get(params);
-
-        if (cached) {
-            return res.status(200).json(cached);
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
         }
 
-        const pageActive = parseInt(page, 10);
-        let offset = 0;
-        if (pageActive === 1 || pageActive === 0) {
-            offset = 0;
-        } else {
-            offset = parseInt(pageActive, 10) * 8 - 8;
-        }
-        const ready = await courses.findAndCountAll({
-            limit: 8,
-            offset,
-        });
-        let next = false;
-        if (ready.count > offset + 8) {
-            next = true;
-        } else {
-            next = false;
-        }
-        cache.set(params, { ...ready, next }, 60 * 20);
-        return res.status(200).json({ ...ready, next });
-    } catch (err) {
-        return res.status(400).json({ err });
-    }
-});
-
-router.post('/create/course', adminMiddleware, async (req, res) => {
-    try {
-        const { title, description, image, time } = req.body;
         await courses.create({
             title,
             description,
             image,
             slug: slug(title, { lower: true }),
+            qt_disciplines: qtDisciplines,
             time,
         });
         return res.status(200).json({ success: 'adicionado com sucesso' });
@@ -167,9 +112,13 @@ router.post('/create/course', adminMiddleware, async (req, res) => {
     }
 });
 
-router.put('/update/course', adminMiddleware, async (req, res) => {
+router.put('/update/course', userMiddleware, async (req, res) => {
     try {
+        const { admin } = req;
         const { title, description, image, id } = req.body;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await courses.update(
             { title, description, image, slug: slug(title, { lower: true }) },
             { where: { id } }
@@ -180,9 +129,13 @@ router.put('/update/course', adminMiddleware, async (req, res) => {
     }
 });
 
-router.delete('/delete/course', adminMiddleware, async (req, res) => {
+router.delete('/delete/course/:id', userMiddleware, async (req, res) => {
     try {
-        const { id } = req.body;
+        const { admin } = req;
+        const { id } = req.params;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await courses.destroy({ where: { id } });
         return res.status(200).json({ success: 'deletado com sucesso!' });
     } catch (err) {
@@ -195,27 +148,13 @@ router.delete('/delete/course', adminMiddleware, async (req, res) => {
 /* Create | Ready | Updade | Delete
 */
 
-router.get('/disciplines', adminMiddleware, async (req, res) => {
+router.post('/disciplines/create', userMiddleware, async (req, res) => {
     try {
-        const params = `discipline:admin`;
-
-        const cached = await cache.get(params);
-
-        if (cached) {
-            return res.status(200).json(cached);
-        }
-
-        const discipline = await disciplines.findAll();
-        cache.set(params, discipline, 60 * 20);
-        return res.status(200).json({ discipline });
-    } catch (err) {
-        return res.status(400).json({ err });
-    }
-});
-
-router.post('/disciplines/create', adminMiddleware, async (req, res) => {
-    try {
+        const { admin } = req;
         const { title, image, qtModules, idCourse } = req.body;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await disciplines.create({
             title,
             slug: slug(title, { lower: true }),
@@ -231,9 +170,13 @@ router.post('/disciplines/create', adminMiddleware, async (req, res) => {
     }
 });
 
-router.put('/disciplines/update/:id', adminMiddleware, async (req, res) => {
+router.put('/disciplines/update/:id', userMiddleware, async (req, res) => {
     try {
+        const { admin } = req;
         const { id } = req.params;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await disciplines.update({ ...req.body }, { where: { id } });
         return res.status(200).json({ success: 'atualizado com sucesso' });
     } catch (err) {
@@ -241,9 +184,13 @@ router.put('/disciplines/update/:id', adminMiddleware, async (req, res) => {
     }
 });
 
-router.delete('/disciplines/delete/:id', adminMiddleware, async (req, res) => {
+router.delete('/disciplines/delete/:id', userMiddleware, async (req, res) => {
     try {
+        const { admin } = req;
         const { id } = req.params;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await disciplines.destroy({ where: { id } });
         return res.status(200).json({ success: 'deletado com sucesso!' });
     } catch (err) {
@@ -256,29 +203,13 @@ router.delete('/disciplines/delete/:id', adminMiddleware, async (req, res) => {
 /* Create | Ready | Update | Delete
 */
 
-router.get('/modules', adminMiddleware, async (req, res) => {
+router.post('/modules/create', userMiddleware, async (req, res) => {
     try {
-        const params = `modules:admin`;
-
-        const cached = await cache.get(params);
-
-        if (cached) {
-            return res.status(200).json(cached);
-        }
-
-        const modulers = await modules.findAll();
-
-        cache.set(params, modulers, 60 * 20);
-
-        return res.status(200).json({ modulers });
-    } catch (err) {
-        return res.status(400).json({ err: 'Ocorreu algum erro' });
-    }
-});
-
-router.post('/modules/create', adminMiddleware, async (req, res) => {
-    try {
+        const { admin } = req;
         const { title, qtClass, qtConcluded, order, idDiscipline } = req.body;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await modules.create({
             title,
             slug: slug(title, { lower: true }),
@@ -293,9 +224,13 @@ router.post('/modules/create', adminMiddleware, async (req, res) => {
     }
 });
 
-router.put('/modules/update/:id', adminMiddleware, async (req, res) => {
+router.put('/modules/update/:id', userMiddleware, async (req, res) => {
     try {
+        const { admin } = req;
         const { id } = req.params;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await modules.update(
             { ...req.body, slug: slug(req.body.title, { lowe: true }) },
             { where: { id } }
@@ -306,9 +241,13 @@ router.put('/modules/update/:id', adminMiddleware, async (req, res) => {
     }
 });
 
-router.delete('/modules/delete/:id', adminMiddleware, async (req, res) => {
+router.delete('/modules/delete/:id', userMiddleware, async (req, res) => {
     try {
+        const { admin } = req;
         const { id } = req.params;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await modules.destroy({ where: { id } });
         return res.status(200).json({ success: 'deleletado com sucesso!' });
     } catch (err) {
@@ -321,27 +260,13 @@ router.delete('/modules/delete/:id', adminMiddleware, async (req, res) => {
 /* Create | Ready | Updade | Delete
 */
 
-router.get('/classes', adminMiddleware, async (req, res) => {
+router.post('/classes/create', userMiddleware, async (req, res) => {
     try {
-        const params = 'classes:admin';
-
-        const cached = await cache.get(params);
-
-        if (cached) {
-            return res.status(200).json(cached);
-        }
-
-        const classe = await classes.findAll();
-        cache.set(params, classe, 60 * 20);
-        return res.status(200).json({ classe });
-    } catch (err) {
-        return res.status(400).json({ err: 'Ocorreu um erro!' });
-    }
-});
-
-router.post('/classes/create', adminMiddleware, async (req, res) => {
-    try {
+        const { admin } = req;
         const { title, url, order, idModule } = req.body;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await classes.create({
             title,
             slug: slug(title, { lower: true }),
@@ -355,9 +280,13 @@ router.post('/classes/create', adminMiddleware, async (req, res) => {
     }
 });
 
-router.put('/classes/update/:id', adminMiddleware, async (req, res) => {
+router.put('/classes/update/:id', userMiddleware, async (req, res) => {
     try {
+        const { admin } = req;
         const { id } = req.params;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await classes.update(
             {
                 ...req.body,
@@ -371,9 +300,13 @@ router.put('/classes/update/:id', adminMiddleware, async (req, res) => {
     }
 });
 
-router.delete('/classes/delete/:id', adminMiddleware, async (req, res) => {
+router.delete('/classes/delete/:id', userMiddleware, async (req, res) => {
     try {
+        const { admin } = req;
         const { id } = req.params;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await classes.destroy({ where: { id } });
         return res.status(200).json({ success: 'deletado com sucesso!' });
     } catch (err) {
@@ -385,27 +318,13 @@ router.delete('/classes/delete/:id', adminMiddleware, async (req, res) => {
 /* Rotas relacionads a materias (pdf, apostilas)
 */
 
-router.get('/materials', adminMiddleware, async (req, res) => {
+router.post('/materials/create', userMiddleware, async (req, res) => {
     try {
-        const params = 'material:admin';
-
-        const cached = await cache.get(params);
-
-        if (cached) {
-            return res.status(200).json(cached);
-        }
-
-        const material = await materials.findAll();
-        cache.set(params, material, 60 * 20);
-        return res.status(200).json({ material });
-    } catch (err) {
-        return res.status(400).json({ err: 'ocorreu algum erro!' });
-    }
-});
-
-router.post('/materials/create', adminMiddleware, async (req, res) => {
-    try {
+        const { admin } = req;
         const { title, url, idModule } = req.body;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await materials.create({ title, url, id_module: idModule });
         return res.status(200).json({ success: 'adiconado com sucesso' });
     } catch (err) {
@@ -413,9 +332,13 @@ router.post('/materials/create', adminMiddleware, async (req, res) => {
     }
 });
 
-router.put('/materials/update/:id', adminMiddleware, async (req, res) => {
+router.put('/materials/update/:id', userMiddleware, async (req, res) => {
     try {
+        const { admin } = req;
         const { id } = req.params;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await materials.update({ ...req.body }, { where: { id } });
         return res.status(200).json({ success: 'atualizado com sucesso' });
     } catch (err) {
@@ -423,9 +346,13 @@ router.put('/materials/update/:id', adminMiddleware, async (req, res) => {
     }
 });
 
-router.delete('/materials/delete/:id', adminMiddleware, async (req, res) => {
+router.delete('/materials/delete/:id', userMiddleware, async (req, res) => {
     try {
+        const { admin } = req;
         const { id } = req.params;
+        if (!admin) {
+            return res.status(400).json({ error: 'you are not an admin' });
+        }
         await materials.destroy({ where: { id } });
         return res.status(200).json({ success: 'deletado com sucesso' });
     } catch (err) {
